--- a/app/src/androidTest/java/com/example/rwothoromo/developers/ExampleInstrumentedTest.java
+++ b/app/src/androidTest/java/com/example/rwothoromo/developers/ExampleInstrumentedTest.java
@@ -7,7 +7,7 @@ import android.support.test.runner.AndroidJUnit4;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -16,11 +16,14 @@ import static org.junit.Assert.*;
  */
 @RunWith(AndroidJUnit4.class)
 public class ExampleInstrumentedTest {
-	@Test
-	public void useAppContext() {
-		// Context of the app under test.
-		Context appContext = InstrumentationRegistry.getTargetContext();
+    /**
+     * Test usage of the AppContext.
+     */
+    @Test
+    public void useAppContext() {
+        // Context of the app under test.
+        Context appContext = InstrumentationRegistry.getTargetContext();
 
-		assertEquals("com.example.rwothoromo.developers", appContext.getPackageName());
-	}
+        assertEquals("com.example.rwothoromo.developers", appContext.getPackageName());
+    }
 }