--- a/app/src/main/java/com/example/rwothoromo/developers/constants/Constants.java
+++ b/app/src/main/java/com/example/rwothoromo/developers/constants/Constants.java
@@ -1,14 +1,18 @@
 package com.example.rwothoromo.developers.constants;
 
 /**
- * This class contains the constants used in this application
+ * This class contains the constants used in this application.
  */
 public class Constants {
 
-	public static final String BASE_URL = "https://api.github.com";
-	public static final String EXTRA_GITHUB_USER_USERNAME = "com.example.rwothoromo.developers.EXTRA_GITHUB_USER_USERNAME";
-	public static final String EXTRA_GITHUB_USER_AVATAR = "com.example.rwothoromo.developers.EXTRA_GITHUB_USER_AVATAR";
-	public static final String EXTRA_GITHUB_USER_URL = "com.example.rwothoromo.developers.EXTRA_GITHUB_USER_URL";
-	public static final String EXTRA_DEVELOPER_LIST_STATE = "com.example.rwothoromo.githubUsers.EXTRA_DEVELOPER_LIST_STATE";
+    public static final String BASE_URL = "https://api.github.com";
+    public static final String EXTRA_GITHUB_USER_USERNAME
+            = "com.example.rwothoromo.developers.EXTRA_GITHUB_USER_USERNAME";
+    public static final String EXTRA_GITHUB_USER_AVATAR
+            = "com.example.rwothoromo.developers.EXTRA_GITHUB_USER_AVATAR";
+    public static final String EXTRA_GITHUB_USER_URL
+            = "com.example.rwothoromo.developers.EXTRA_GITHUB_USER_URL";
+    public static final String EXTRA_DEVELOPER_LIST_STATE
+            = "com.example.rwothoromo.githubUsers.EXTRA_DEVELOPER_LIST_STATE";
 
 }