--- a/app/src/main/java/com/example/rwothoromo/developers/model/GithubUser.java
+++ b/app/src/main/java/com/example/rwothoromo/developers/model/GithubUser.java
@@ -6,89 +6,105 @@ import android.os.Parcelable;
 import com.google.gson.annotations.SerializedName;
 
 /**
- * This class represents the GitHub User model
+ * This class represents the GitHub User model.
  */
 public class GithubUser implements Parcelable {
-	public static final Creator<GithubUser> CREATOR
-			= new Creator<GithubUser>() {
-		public GithubUser createFromParcel(Parcel in) {
-			return new GithubUser(in);
-		}
+    public static final Creator<GithubUser> CREATOR
+            = new Creator<GithubUser>() {
+        public GithubUser createFromParcel(Parcel in) {
+            return new GithubUser(in);
+        }
 
-		public GithubUser[] newArray(int size) {
-			return new GithubUser[size];
-		}
-	};
-	private int id;
-	@SerializedName("login")
-	private String username;
-	@SerializedName("avatar_url")
-	private String avatarUrl;
-	@SerializedName("html_url")
-	private String htmlUrl;
-	@SerializedName("organizations_url")
-	private String organizationsUrl;
+        public GithubUser[] newArray(int size) {
+            return new GithubUser[size];
+        }
+    };
+    private int id;
+    @SerializedName("login")
+    private String username;
+    @SerializedName("avatar_url")
+    private String avatarUrl;
+    @SerializedName("html_url")
+    private String htmlUrl;
+    @SerializedName("organizations_url")
+    private String organizationsUrl;
 
-	private GithubUser(Parcel in) {
-		id = in.readInt();
-		username = in.readString();
-		avatarUrl = in.readString();
-		htmlUrl = in.readString();
-	}
+    /**
+     * Constructor for the GithubUser class.
+     *
+     * @param in a Parcelable GitHub user object
+     */
+    private GithubUser(Parcel in) {
+        id = in.readInt();
+        username = in.readString();
+        avatarUrl = in.readString();
+        htmlUrl = in.readString();
+    }
 
-	/**
-	 * Returns the GitHub user id
-	 *
-	 * @return id
-	 */
-	public int getId() {
-		return id;
-	}
+    /**
+     * Returns the GitHub user id.
+     *
+     * @return id
+     */
+    public int getId() {
+        return id;
+    }
 
-	/**
-	 * Returns the GitHub username
-	 *
-	 * @return username
-	 */
-	public String getUsername() {
-		return username;
-	}
+    /**
+     * Returns the GitHub username.
+     *
+     * @return username
+     */
+    public String getUsername() {
+        return username;
+    }
 
-	/**
-	 * Returns the GitHub user profile image URL
-	 *
-	 * @return avatarUrl
-	 */
-	public String getAvatarUrl() {
-		return avatarUrl;
-	}
+    /**
+     * Returns the GitHub user profile image URL.
+     *
+     * @return avatarUrl
+     */
+    public String getAvatarUrl() {
+        return avatarUrl;
+    }
 
-	/**
-	 * Returns the GitHub user profile URL
-	 *
-	 * @return htmlUrl
-	 */
-	public String getHtmlUrl() {
-		return htmlUrl;
-	}
+    /**
+     * Returns the GitHub user profile URL.
+     *
+     * @return htmlUrl
+     */
+    public String getHtmlUrl() {
+        return htmlUrl;
+    }
 
-	/**
-	 * Returns the GitHub user organizations URL
-	 *
-	 * @return organizationsUrl
-	 */
-	public String getOrganizationsUrl() {
-		return organizationsUrl;
-	}
+    /**
+     * Returns the GitHub user organizations URL.
+     *
+     * @return organizationsUrl
+     */
+    public String getOrganizationsUrl() {
+        return organizationsUrl;
+    }
 
-	public int describeContents() {
-		return 0;
-	}
+    /**
+     * Child classes can return different values with this method.
+     *
+     * @return integer
+     */
+    public int describeContents() {
+        return 0;
+    }
 
-	public void writeToParcel(Parcel out, int flags) {
-		out.writeInt(id);
-		out.writeString(username);
-		out.writeString(avatarUrl);
-		out.writeString(htmlUrl);
-	}
+    /**
+     * Write data to the Parcel.
+     *
+     * @param out a Parcel
+     * @param flags flags
+     */
+    public void writeToParcel(Parcel out, int flags) {
+        out.writeInt(id);
+        out.writeString(username);
+        out.writeString(avatarUrl);
+        out.writeString(htmlUrl);
+    }
 }