--- a/app/src/main/java/com/example/rwothoromo/developers/service/GithubService.java
+++ b/app/src/main/java/com/example/rwothoromo/developers/service/GithubService.java
@@ -6,27 +6,27 @@ import retrofit2.converter.gson.GsonConverterFactory;
 import static com.example.rwothoromo.developers.constants.Constants.BASE_URL;
 
 /**
- * This class is for the GitHub service
- * It helps in sending out network requests to the GitHub API
+ * This class is for the GitHub service.
+ * It helps in sending out network requests to the GitHub API.
  */
 public class GithubService {
 
-	private Retrofit retrofit = null;
+    private Retrofit retrofit = null;
 
-	/**
-	 * Returns a new instance of the GithubAPI interface
-	 *
-	 * @return GithubAPI interface
-	 */
-	public GithubAPI getAPI() {
+    /**
+     * Returns a new instance of the GithubAPI interface.
+     *
+     * @return GithubAPI interface
+     */
+    public GithubAPI getAPI() {
 
-		if (retrofit == null) {
-			retrofit = new Retrofit.Builder()
-					.baseUrl(BASE_URL)
-					.addConverterFactory(GsonConverterFactory.create())
-					.build();
-		}
+        if (retrofit == null) {
+            retrofit = new Retrofit.Builder()
+                    .baseUrl(BASE_URL)
+                    .addConverterFactory(GsonConverterFactory.create())
+                    .build();
+        }
 
-		return retrofit.create(GithubAPI.class);
-	}
+        return retrofit.create(GithubAPI.class);
+    }
 }