--- a/app/src/test/java/com/example/rwothoromo/developers/ExampleUnitTest.java
+++ b/app/src/test/java/com/example/rwothoromo/developers/ExampleUnitTest.java
@@ -2,7 +2,7 @@ package com.example.rwothoromo.developers;
 
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
 /**
  * Example local unit test, which will execute on the development machine (host).
@@ -10,8 +10,11 @@ import static org.junit.Assert.*;
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
 public class ExampleUnitTest {
-	@Test
-	public void additionIsCorrect() {
-		assertEquals(4, 2 + 2);
-	}
+    /**
+     * Test that addition works correctly.
+     */
+    @Test
+    public void additionIsCorrect() {
+        assertEquals(4, 2 + 2);
+    }
 }