--- a/gradle.properties
+++ b/gradle.properties
@@ -11,3 +11,15 @@ org.gradle.jvmargs=-Xmx1536m
 # This option should only be used with decoupled projects. More details, visit
 # http://www.gradle.org/docs/current/userguide/multi_project_builds.html#sec:decoupled_projects
 # org.gradle.parallel=true
+
+# Gradle build cache for faster builds
+org.gradle.caching=true
+
+# Enable reusing things in subsequent builds
+org.gradle.daemon=true
+
+# Allow Gradle to build modules in parallel; only useful if the modules don't depend on each other
+org.gradle.parallel=true
+
+# Configures only needed modules of the project
+org.gradle.configureondemand=true