--- a/quality/quality.gradle
+++ b/quality/quality.gradle
@@ -29,7 +29,8 @@ task checkstyle(type: Checkstyle) {
     source 'src'
     include '**/*.java'
     exclude '**/gen/**',
-            '**/androidTest/**'
+            '**/androidTest/**',
+            '**/test/**'
     classpath = files()
     ignoreFailures = false
 }