… app/src/main/java/com/example/rwothoromo/developers/view/MainActivity.java
… quality/checkstyle/checkstyle.xml