#### What does this PR do?
Speeds up Gradle builds.

#### Description of Task to be completed?
- Add commands to cache Gradle builds for faster future builds
- Add a `pre-commit.sh` to allow Gradle tasks to run before git commits
- Implement changes requested after running `./gradlew check`

#### How should this be manually tested?
Run `./gradle --build-cache clean assemble` in your terminal and note the build time.
Run the command again and note the build time; it is much shorter because it utilized the cache created.

#### Any background context you want to provide?
The builds were not being cached initially.

#### What are the relevant pivotal tracker stories?
[#158787482](https://www.pivotaltracker.com/story/show/158787482)

#### Screenshots (if appropriate)?
1. Before caching (49 seconds).
<img width="1440" alt="initial gradle build - 49s" src="https://user-images.githubusercontent.com/7848682/44217315-d6b67380-a17f-11e8-9d88-e1486dfb493a.png">

2. After caching (18 seconds).
<img width="1440" alt="gradle build using cache - 18s" src="https://user-images.githubusercontent.com/7848682/44217325-df0eae80-a17f-11e8-8ca9-c40d76eaf52d.png">